+++ internal/digraph/step_context.go
@@ -64,7 +64,7 @@ func (c StepContext) MailerConfig() (mailer.Config, error) {
 func (c StepContext) EvalString(s string, opts ...cmdutil.EvalOption) (string, error) {
 	opts = append(opts, cmdutil.WithVariables(c.envs))
 	opts = append(opts, cmdutil.WithVariables(c.outputVariables.Variables()))
-	return cmdutil.EvalString(s, opts...)
+	return cmdutil.EvalString(c.ctx, s, opts...)
 }
 
 func (c StepContext) EvalBool(value any) (bool, error) {
@@ -107,6 +107,6 @@ func IsStepContext(ctx context.Context) bool {
 type stepCtxKey struct{}
 
 func EvalStringFields[T any](stepContext StepContext, obj T) (T, error) {
-	return cmdutil.EvalStringFields(obj,
+	return cmdutil.EvalStringFields(stepContext.ctx, obj,
 		cmdutil.WithVariables(stepContext.outputVariables.Variables()))
 }